Project Requirements Document Section 1
Entire format has been reconstructed to comply with IEEE standards as suggested by Winters. I've completed my first pass on the first third of the document and have outlined sections 2 and 3. 

Once we agree on body text and general formatting we can proceed to convert into a suitable tex and send to Chris.