Revisions made to section 3
Changes have been completed and the final section was mostly intact. I've noticed the source of our potential problem for compiling. School servers appear to naturally lack a few packages we're trying to include. I got a pdf done by locally installing a compiler but we'll need to correct this prior to final submission.

Lastly, figures 1 and 2 come out blurred when compared to the rest of the text. Another point to change.