Issues with compile resolved
Changes made to:
Makefile
design_doc.tex

The method of calling italics was causing compiler errors on the flip server, adjusted how those are called as well as swapping out our makefile that would fail to compile on the first attempt but succeed on the second. Not sure what caused that.